# specs/001-gs-control-center/contracts/api.yaml
openapi: 3.1.0
info:
  title: GS Control Center API
  description: Centralized Advertising Block Management System for UCS Guest Screen
  version: 1.0.0
servers:
  - url: /api/v1
    description: Central Server API (Host 10.0.0.111)
paths:
...
  /media:
    get:
      summary: List media assets
      parameters:
        - name: media_type
          in: query
          schema: { type: string, enum: [IMAGE, VIDEO] }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MediaAsset'
  /media/upload:
    post:
      summary: Upload image or video asset to MinIO
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MediaAsset'
  /media/{id}:
    patch:
      summary: Rename media asset display name (OCC protected)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [original_name, version]
              properties:
                original_name: { type: string, maxLength: 255 }
                version: { type: integer }
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MediaAsset'
        '409':
          description: Optimistic lock conflict (media was modified by another operator)
    delete:
      summary: Delete media asset (safe deletion with dependency protection & S3 refcount)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204':
          description: Media asset deleted
        '409':
          description: Media asset is in use by advertising blocks (deletion rejected)
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail: { type: string }
                  dependent_templates:
                    type: array
                    items:
                      type: object
                      properties:
                        id: { type: string, format: uuid }
                        name: { type: string }
                        area: { type: string }
  /media/{id}/replace:
    post:
      summary: In-place media binary replacement (recomputes SHA-256, preserves template references)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, version]
              properties:
                file:
                  type: string
                  format: binary
                version:
                  type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MediaAsset'
        '409':
          description: Optimistic lock conflict (version mismatch)
  /media/{id}/usage:
    get:
      summary: Inspect templates referencing this media asset
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  media_id: { type: string, format: uuid }
                  usage_count: { type: integer }
                  templates:
                    type: array
                    items:
                      type: object
                      properties:
                        id: { type: string, format: uuid }
                        name: { type: string }
                        area: { type: string }
                        display_mode: { type: string }
...
components:
  schemas:
...
    MediaAsset:
      type: object
      properties:
        id: { type: string, format: uuid }
        original_name: { type: string }
        stored_name: { type: string }
        sha256: { type: string }
        mime_type: { type: string }
        media_type: { type: string, enum: [IMAGE, VIDEO] }
        file_size_bytes: { type: integer }
        width: { type: integer, nullable: true }
        height: { type: integer, nullable: true }
        version: { type: integer, description: OCC version counter }
        preview_url: { type: string }
...